通用异常处理-增加评论-优化逻辑判断

+++ src/main/java/com/community/controller/QuestionController.java
@@ -18,7 +18,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("question/{id}")
-    public String getQuestion(@PathVariable(name = "id") Integer id,
+    public String getQuestion(@PathVariable(name = "id") Long id,
                               Model model){
 
         questionService.addViewCount(id);
